requirements.txt
sympy
Pillow
numpy
mpmath
scikit-learn
scipy